requirements.txt
GitPython==3.1.44
PyGithub==2.5.0
PyYAML<=6.0.2
python-hcl2==7.0.1
prettytable==3.16.0